Main code of other processes
The following class represents the main code performed by the other processes when the distributed write and distributed read are coming.
The generic process will wait a notification for the following operations:
-WRITE, when the process receive this notification, the distributed write will be performed
-READ, when the process receive this notification, the distributed read will be performed
-TERMINATE, when the process receive this notification, it must terminate its execution

Similar to the code executed by the master process, the following class contains the counterpart of the MPI collective calls for the generic node of the distributed system.

DataBlock data structure, in this case store the actual data rather than just pointers as the master process does.
The data to be stored is distributed by the master process.