Created the DataBlockManager class
This class performs the initialization of data blocks that will contain valid pointers to the memory areas of other processes in the distributed system.
Each block is initialized with the rank of the process that will store it, a progressive number and an offset.
The pointer to the actual data will be received after distributing the data