Created request packages
The following packages are used for interactions and data exchanges between processes.

When the master process has to notify to other processes the which operation has to be performed, it will send a "RequestPacket" will send a "Request Packet" specifying the type of request in the "RequestType" field

The "IORequestPacket" is the packet that the master process sends to share the distributed operation data

The "PointerPacket" is the packet that contains the valid memory address where the data was actually stored.