Master process creating the mountpoint directory
Now, the master process will create the mountpoint directory at the given path. It also check if the directory already exists.
The master process also delete the mountpoint directory at the end of the programm.
Future implementations will allow to user to choose if the mountpoint directory will be kept in the file system.